refactor: renamed preset.py to assignment_config.py
