+++ connectors/models/assignment_config.py
